requirements.txt
# Core dependencies
unstructured[all-docs]==0.15.13
sentence-transformers==3.3.1
qdrant-client==1.12.1
pyyaml==6.0.2
tqdm==4.67.1
numpy==1.26.4
requests==2.32.3

# Optional: For better sentence chunking
# nltk==3.9.1
# spacy==3.8.2
